test: test for chunk verify cycle calculation
